Redo how port markers work
Port markers are now shown by default (rather than toggle, but having
lost their hotkey). They also now reflect whether a connection points
in or out or is bidirectional.

This is a first-pass though, in that the cases of un-formed or
half-formed connections aren't currently handled well--they just don't
show any indicator at all, leaving you to guess where on the outside the
connection point is.